Move meta info into separate file
